Refactor EventProps interface to include duration as a number
