--- src/types/events/EventProps.ts
@@ -1,9 +1,0 @@
-export interface EventProps {
-  id: number
-  name: string
-  location: string
-  dateAndTime: string
-  duration: 1 | 2 | 3 | 4 | 5 | 6
-  description: string
-  customUrl: string
-}
